requirements.txt
streamlit
openai
PyPDF2
python-dotenv
spacy
jinja2
pdfkit
